requirements.txt
torch==2.1.0
transformers==4.34.0
weaviate_client==3.24.2
